+++ lib/src/clojure_lsp/feature/java_interop.clj
@@ -1,5 +1,6 @@
 (ns clojure-lsp.feature.java-interop
   (:require
+   [babashka.fs :as fs]
    [clojure-lsp.config :as config]
    [clojure-lsp.db :as db]
    [clojure-lsp.http :as http]
@@ -14,6 +15,7 @@
    (java.net
      URL
      JarURLConnection)
+   (java.nio.file Path)
    (java.util.zip ZipInputStream)
    (org.benf.cfr.reader.api
      CfrDriver
@@ -21,11 +23,13 @@
 
 (set! *warn-on-reflection* true)
 
+(def ^:private logger-tag (shared/colorize "[Java]" :bright-yellow))
+
 (defn ^:private decompile! [^File class-file dest-path db]
   (let [cache-path (config/cache-file db)
         decompiled-file (io/file cache-path "java" "decompiled")
         class-path (.getCanonicalPath class-file)
-        _ (logger/info (format "[JDK] Decompiling java class %s" class-path))
+        _ (logger/info logger-tag (format "Decompiling java class %s" class-path))
         driver ^CfrDriver (.. (CfrDriver$Builder.)
                               ;; CFR stout is not reliable, prefer output to file so we can read later
                               (withOptions {"outputdir" (.getCanonicalPath decompiled-file)})
@@ -35,13 +39,13 @@
       (with-out-str
         (.analyse driver [class-path]))
       (when-not (string/blank? (str err-sym))
-        (logger/warn "Non-fatal error from CFR:" (str err-sym))))
+        (logger/warn logger-tag "Non-fatal error from CFR:" (str err-sym))))
     (io/file decompiled-file dest-path)))
 
 (defn ^:private copy-class-file [uri entry stream db]
   (let [cache-path (config/cache-file db)
         dest-file (io/file cache-path "java" "classes" (str entry))]
-    (logger/info (format "[JDK] Copying class URI %s to %s" uri dest-file))
+    (logger/info logger-tag (format "Copying class URI %s to %s" uri dest-file))
     (io/make-parents dest-file)
     (io/copy stream dest-file)
     dest-file))
@@ -68,6 +72,33 @@
 (defn read-content! [uri components]
   (slurp (uri->translated-file uri components)))
 
+(def ^:private jdk-source-zip-filename "src.zip")
+
+(defn ^:private jdk-path->jdk-source-paths [^File root-path]
+  (when root-path
+    (let [parent (.getParentFile root-path)]
+      [(io/file root-path jdk-source-zip-filename)
+       (io/file root-path "lib" jdk-source-zip-filename)
+       (io/file parent jdk-source-zip-filename)
+       (io/file parent "lib" jdk-source-zip-filename)])))
+
+(defn ^:private find-local-jdk-source []
+  (let [java-home (some-> (or (config/get-property "java.home")
+                              (config/get-env "JAVA_HOME")) io/file)
+        paths (concat []
+                      (jdk-path->jdk-source-paths java-home)
+                      (->> (fs/which-all "java")
+                           (map (comp
+                                  jdk-path->jdk-source-paths
+                                  #(some-> ^File % .getParentFile .getParentFile)
+                                  #(.toFile ^Path %)
+                                  fs/real-path))
+                           flatten))]
+    (some->> paths
+             (filter #(.canRead ^File %))
+             first))
+  nil)
+
 (def ^:private default-jdk-source-uri
   "https://raw.githubusercontent.com/clojure-lsp/jdk-source/main/openjdk-19/reduced/source.zip")
 
@@ -75,15 +106,14 @@
   [^String jdk-download-url
    ^File dest-jdk-file]
   (io/make-parents dest-jdk-file)
-  (logger/info "[JDK] Downloading JDK source...")
   (shared/logging-time
-    "[JDK] Downloading JDK source took %s secs."
+    (str logger-tag " Downloading JDK source took %s secs.")
     (try
       (let [{:keys [body content-type status error]} (http/request! jdk-download-url)]
         (if (or error
                 (not= 200 status))
-          (logger/error "[JDK] Could not download JDK source." error)
-          ;; TODO handle mode common downloadable content-types
+          (logger/error logger-tag "Could not download JDK source." error)
+          ;; TODO handle more common downloadable content-types
           (case content-type
             "application/zip"
             (with-open [stream (ZipInputStream. body)]
@@ -99,13 +129,13 @@
                           (.mkdirs parentDir))
                         (clojure.java.io/copy stream saveFile)))
                     (recur (.getNextEntry stream))))))
-            (logger/error "[JDK] Could not download JDK source, unknown content-type" content-type))))
+            (logger/error logger-tag "Could not download JDK source, unknown content-type" content-type))))
       (catch Exception e
-        (logger/error "[JDK] Error Downloading JDK source." e)))))
+        (logger/error logger-tag "Error Downloading JDK source." e)))))
 
 (defn ^:private analyze-jdk-source! [path db]
   (let [result (shared/logging-time
-                 "[JDK] Analyzing JDK source with clj-kondo took %s secs."
+                 (str logger-tag " Analyzing JDK source with clj-kondo took %s secs.")
                  (lsp.kondo/run-kondo-on-jdk-source! path))
         kondo-analysis (select-keys (:analysis result) [:java-class-definitions])
         analysis (->> kondo-analysis
@@ -113,31 +143,60 @@
                       (group-by :filename))]
     (loop [state-db @db]
       (when-not (compare-and-set! db state-db (update state-db :analysis merge analysis))
-        (logger/warn "Analyzis outdated from java analysis, trying again...")
+        (logger/warn logger-tag "Analyzis outdated from java analysis, trying again...")
         (recur @db)))
     (-> (shared/uri->path (:project-root-uri @db))
         (db/read-cache db)
         (update :analysis merge analysis)
         (db/upsert-cache! db))))
 
-(defn ^:private download-jdk-source
-  [jdk-source-uri db]
-  (let [global-cache-dir (config/global-lsp-cache-dir)
-        dest-jdk-file (io/file global-cache-dir "jdk")
-        dest-jdk-result-file (io/file dest-jdk-file "successfully-downloaded")]
-    (if (shared/file-exists? dest-jdk-result-file)
-      (logger/info "[JDK] Skipping downloading JDK source, already found at" dest-jdk-file)
-      (do
-        (download-jdk! jdk-source-uri dest-jdk-file)
-        (spit dest-jdk-result-file "1")))
-    (.getCanonicalPath dest-jdk-file) db))
+(defn retrieve-jdk-source-and-analyze!
+  "Find JDK source and analyze it with clj-kondo for java class definitions.
 
-(defn retrieve-and-analyze! [{:keys [db]}]
-  (let [jdk-source-uri (settings/get db [:java :jdk-source-uri] default-jdk-source-uri)
-        download-jdk-source? (settings/get db [:java :download-jdk-source?] false)]
-    (when-let [jdk-source-uri' (or (and (shared/plain-uri? jdk-source-uri)
-                                        (shared/filename->uri jdk-source-uri db))
-                                   (when download-jdk-source? (download-jdk-source jdk-source-uri db)))]
-      (logger/info (str "[JDK] Analyzing JDK source found at " jdk-source-uri "..."))
-      (analyze-jdk-source! jdk-source-uri' db)
-      (logger/info "[JDK] Source analyzed successfully."))))
+  If user has custom :jdk-source-uri, we don't try to find JDK source local
+  or download it, we make sure zip is extracted otherwize extract it.
+
+  If no custom URI set, we try to find a local JDK source on Java
+  instalation which is present most of the time and extract to global
+  cache dir.
+
+  Otherwise, we download default OpenJDK source if setting is enabled."
+  [{:keys [db]}]
+  (let [global-cache-dir (config/global-lsp-cache-dir)
+        jdk-dir-file (io/file global-cache-dir "jdk")
+        jdk-result-file (io/file jdk-dir-file "result")
+        installed-jdk-source-uri (and (shared/file-exists? jdk-result-file)
+                                      (slurp jdk-result-file))
+        user-jdk-source-uri (settings/get db [:java :jdk-source-uri])
+        jdk-source-uri (or user-jdk-source-uri default-jdk-source-uri)]
+
+    (if-not installed-jdk-source-uri
+      (if-let [local-jdk-source-zip (and (not user-jdk-source-uri)
+                                         (find-local-jdk-source))]
+        (do
+          (logger/info logger-tag "Found local JDK source zip, extracting to global LSP cache dir...")
+          (fs/unzip local-jdk-source-zip jdk-dir-file {:replace-existing true})
+          (spit jdk-result-file (.getCanonicalPath ^File local-jdk-source-zip)))
+        (do
+          (logger/info logger-tag "Local JDK source not found.")
+          (if-let [local-jdk-uri (and (shared/plain-uri? jdk-source-uri)
+                                      (shared/filename->uri jdk-source-uri db))]
+            (do
+              (logger/info logger-tag "Found local JDK source URI, extracting to global LSP cache dir...")
+              (fs/unzip local-jdk-uri jdk-dir-file {:replace-existing true})
+              (spit jdk-result-file local-jdk-uri))
+            (if (settings/get db [:java :download-jdk-source?] false)
+              (do
+                (logger/info logger-tag "Downloading JDK source to global LSP cache dir...")
+                (download-jdk! jdk-source-uri jdk-dir-file)
+                (spit jdk-result-file jdk-source-uri))
+              (logger/warn logger-tag "Skipping download JDK source, setting `:java :download-jdk-source?` is disabled.")))))
+      (logger/info logger-tag "JDK source already present on global LSP cache dir."))
+
+    (if (or installed-jdk-source-uri
+            (and (shared/file-exists? jdk-result-file)
+                 (slurp jdk-result-file)))
+      (do
+        (analyze-jdk-source! (.getCanonicalPath jdk-dir-file) db)
+        (logger/info logger-tag "JDK Source analyzed successfully."))
+      (logger/warn logger-tag "JDK Source not found, skipping java analysis."))))